+++ backend/auth/requests/credential.py
@@ -3,20 +3,20 @@ from dataclasses import dataclass
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework import status
-from common.views.common.requests.base import BaseRequest, Headers
+from common.requests.base import BaseRequest, Headers
 
 
 @dataclass(frozen=True)  # TODO: pydantic validation
-class EmailTokenRequest(BaseRequest):
+class CredentialRequest(BaseRequest):
     email: str
-    token: str
+    password: str
 
     @staticmethod
-    def _init(request: Request) -> 'EmailTokenRequest':
+    def _init(request: Request) -> 'CredentialRequest':
         request_data = dict(request.data)
-        return EmailTokenRequest(
+        return CredentialRequest(
             email=str(request_data.get('email')),
-            token=str(request_data.get('token')),
+            password=str(request_data.get('password')),
             user=str(request.user),
             content_type=str(request.content_type),
             auth=request.auth,
@@ -24,11 +24,11 @@ class EmailTokenRequest(BaseRequest):
         )
 
     @staticmethod
-    def init(request_data: Request) -> 'EmailTokenRequest' | Response:
+    def init(request_data: Request) -> 'CredentialRequest' | Response:
         try:
-            return EmailTokenRequest._init(request_data)
+            return CredentialRequest._init(request_data)
         except TypeError:
             return Response(
-                {"text": "Request data is not valid. Please provide 'email' and 'token' fields."},
+                {"text": "Request data is not valid. Please provide 'email' and 'password' fields."},
                 status=status.HTTP_400_BAD_REQUEST,
             )