+++ backend/auth/requests/email_token.py
@@ -3,20 +3,20 @@ from dataclasses import dataclass
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework import status
-from common.views.common.requests.base import BaseRequest, Headers
+from common.requests.base import BaseRequest, Headers
 
 
 @dataclass(frozen=True)  # TODO: pydantic validation
-class CredentialRequest(BaseRequest):
+class EmailTokenRequest(BaseRequest):
     email: str
-    password: str
+    token: str
 
     @staticmethod
-    def _init(request: Request) -> 'CredentialRequest':
+    def _init(request: Request) -> 'EmailTokenRequest':
         request_data = dict(request.data)
-        return CredentialRequest(
+        return EmailTokenRequest(
             email=str(request_data.get('email')),
-            password=str(request_data.get('password')),
+            token=str(request_data.get('token')),
             user=str(request.user),
             content_type=str(request.content_type),
             auth=request.auth,
@@ -24,11 +24,11 @@ class CredentialRequest(BaseRequest):
         )
 
     @staticmethod
-    def init(request_data: Request) -> 'CredentialRequest' | Response:
+    def init(request_data: Request) -> 'EmailTokenRequest' | Response:
         try:
-            return CredentialRequest._init(request_data)
+            return EmailTokenRequest._init(request_data)
         except TypeError:
             return Response(
-                {"text": "Request data is not valid. Please provide 'email' and 'password' fields."},
+                {"text": "Request data is not valid. Please provide 'email' and 'token' fields."},
                 status=status.HTTP_400_BAD_REQUEST,
             )