+++ backend/auth/views/check_auth.py
@@ -5,7 +5,7 @@ from rest_framework import viewsets, status
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework.permissions import AllowAny
-from auth.views.common.requests.email_token import EmailTokenRequest
+from auth.requests.email_token import EmailTokenRequest
 from rest_framework.authtoken.models import Token