+++ backend/budget/requests/budget.py
@@ -1,9 +1,7 @@
 from __future__ import annotations
 from dataclasses import dataclass
 from rest_framework.request import Request
-from rest_framework.response import Response
-from rest_framework import status
-from common.views.common.requests.base import BaseRequest, Headers
+from common.requests.base import BaseRequest, Headers
 
 
 @dataclass(frozen=True)  # TODO: pydantic validation
@@ -12,7 +10,7 @@ class BudgetRequest(BaseRequest):
     total_budget: int
 
     @staticmethod
-    def _init(request: Request) -> 'BudgetRequest':
+    def init(request: Request) -> 'BudgetRequest':
         request_data = dict(request.data)
         return BudgetRequest(
             name=str(request_data.get('name')),
@@ -22,13 +20,3 @@ class BudgetRequest(BaseRequest):
             auth=request.auth,
             headers=Headers.init(request.headers),
         )
-
-    @staticmethod
-    def init(request_data: Request) -> 'BudgetRequest' | Response:
-        try:
-            return BudgetRequest._init(request_data)
-        except TypeError:
-            return Response(
-                {"text": "Request data is not valid. Please provide 'name' and 'total_budget' fields."},
-                status=status.HTTP_400_BAD_REQUEST,
-            )