+++ backend/common/responses/abstract_response.py
@@ -1,0 +1,12 @@
+from abc import ABC, abstractmethod
+from dataclasses import dataclass
+from rest_framework.response import Response
+
+
+@dataclass
+class AbstractResponse(ABC):
+    
+    @abstractmethod
+    def response(self) -> Response:
+        """ """
+