+++ docs/CHANGELOG.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [[0.6.0-alpha]] - 2022-03-25
+
+###### _([diff: 0.5.0-0.6.0-alpha])_
+
+### Added
+
+- In-progress: Creating Request Handlers
+
+### Changed
+
+- Refactorizations on backend:
+    - Created new directories `requests` and `responses`
+    
+
+### Fixed
+
+- Hiding Modal after user finishes adding a new budget.
+
+
 ## [[0.5.0]] - 2022-03-25
 
 ###### _([diff: 0.4.6-0.5.0])_
@@ -289,6 +308,7 @@ Connected Frontend & Backend via webpack_loader.
     -   I know I should implement them right away and have all of that sort of things already done, but I _really_ want to start some dev coding 🐈
     -   ... still got to setup frontend-backend, so I'll prioritize that at this moment.
 
+[diff: 0.5.0-0.6.0-alpha]: https://github.com/Luzkan/FamilyBudget/compare/0.5.0...0.6.0-alpha
 [diff: 0.4.6-0.5.0]: https://github.com/Luzkan/FamilyBudget/compare/0.4.6...0.5.0
 [diff: 0.4.5-0.4.6]: https://github.com/Luzkan/FamilyBudget/compare/0.4.5...0.4.6
 [diff: 0.4.4-0.4.5]: https://github.com/Luzkan/FamilyBudget/compare/0.4.4...0.4.5
@@ -309,6 +329,7 @@ Connected Frontend & Backend via webpack_loader.
 [diff: 0.1.1-0.2.0]: https://github.com/Luzkan/FamilyBudget/compare/0.1.1...0.2.0
 [diff: 0.1.0-0.1.1]: https://github.com/Luzkan/FamilyBudget/compare/0.1.0...0.1.1
 [diff: 0.0.0-0.1.0]: https://github.com/Luzkan/FamilyBudget/compare/0.0.0...0.1.0
+[0.6.0-alpha]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.6.0-alpha
 [0.5.0]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.5.0
 [0.4.6]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.4.6
 [0.4.5]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.4.5